Began refactoring code
