+++ src/main.py
@@ -2,7 +2,9 @@ import os
 import sys
 import shutil
 import json
-from dataclasses import dataclass
+from file_management import FileDirectory, FileManagement
+from game_data import GameData, GameDataManager
+
 
 DEFAULT_MANIFESTS_PATH: str = "C:\\ProgramData\\Epic\\EpicGamesLauncher\\Data\\Manifests"
 MANIFEST_BACKUP_FOLDER_NAME: str = "_MANIFEST_BACKUPS"
@@ -14,134 +16,14 @@ STAGING_FOLDER_NAME: str = "bps"
 
 LAUNCHER_MANIFEST_FILE_TYPE: str = ".item"
 SUPPORTED_LAUNCHER_MANIFEST_VERSIONS: list[int] = [0,]
-
-@dataclass
-class FileDirectory:
-    name: str # Name of file or folder.
-    path: str # Path to file or folder.
-
-    def get_name_raw(self) -> str:
-        return os.path.splitext(self.name)[0]
     
-    def get_extension(self) -> str:
-        return os.path.splitext(self.name)[1]
-
-@dataclass
-class GameData:
-    game_folder: FileDirectory
-    manifest_folder: FileDirectory
-    manifest_file_list: list[FileDirectory]
-
 
 def print_line_separator() -> None:
     print('-' * 40)
 
-def assert_path_exists(path: str, hint: str = "") -> None:
-    if os.path.exists(path) == False:
-        print(f"ERROR!: Path does not exist! \"{path}\"")
-
-        if len(hint) != 0:
-            print(f"ERROR!: {hint}")
-
-        sys.exit(1)
-
-def try_create_dir(path: str) -> None:
-    if os.path.exists(path) == False:
-        print(f"INFO: Creating folder: {path}")
-        os.mkdir(path)
-
 def yes_no_prompt(prompt: str) -> bool:
     option: str = input(f"{prompt} (y/n): ")
     return len(option) != 0 and str.upper(option[0]) == 'Y'
-
-def is_valid_launcher_manifest_file(entry: os.DirEntry) -> bool:
-    return (
-        entry.is_file() 
-        and LAUNCHER_MANIFEST_FILE_TYPE in entry.name
-    )
-
-def is_valid_game_manifest_file(entry: os.DirEntry) -> bool:
-    return (
-        entry.is_file() 
-        and GAME_MANIFEST_FILE_TYPE in entry.name
-    )
-
-def is_valid_game_folder(entry: os.DirEntry) -> bool:
-    return (
-        entry.is_dir() 
-        and entry.name != MANIFEST_BACKUP_FOLDER_NAME
-        and os.path.exists(os.path.join(entry.path, GAME_MANIFEST_FOLDER_NAME))
-    )
-
-def get_launcher_manifest_files(launcher_manifests_path: str) -> list[FileDirectory]:
-    
-    launcher_manifest_file_list: list[FileDirectory] = []
-
-    for manifest_entry in os.scandir(launcher_manifests_path):
-        if is_valid_launcher_manifest_file(manifest_entry):
-            launcher_manifest_file_list.append(FileDirectory(manifest_entry.name, manifest_entry.path))
-    
-    return launcher_manifest_file_list
-
-def get_game_data_list(game_data_path: str):
-    
-    game_data_list: list[GameData] = []
-    
-    for game_entry in os.scandir(game_data_path):
-        
-        game_folder: FileDirectory = None
-        manifest_folder: FileDirectory = None
-        manifest_file_list: list[FileDirectory] = []
-
-        if is_valid_game_folder(game_entry) == False:
-            print(f"WARNING!: Skipping \"{game_entry.name}/\" as it is not a valid game folder.")
-            continue
-
-        manifest_folder_path = os.path.join(game_entry.path, GAME_MANIFEST_FOLDER_NAME)
-
-        for manifest_entry in os.scandir(manifest_folder_path):
-            if is_valid_game_manifest_file(manifest_entry):
-                manifest_file_list.append(FileDirectory(manifest_entry.name, manifest_entry.path))
-        
-        if len(manifest_file_list) == 0:
-            print(f"WARNING!: Skipping \"{game_entry.name}/.egstore/\" as it is missing a manifest file. (May be an incomplete installation).")
-            continue
-        
-        game_folder = FileDirectory(game_entry.name, game_entry.path)
-        manifest_folder = FileDirectory(GAME_MANIFEST_FOLDER_NAME, manifest_folder_path)
-
-        print(f"INFO: Adding \"{game_entry.name}\"")
-        game_data_list.append(GameData(game_folder, manifest_folder, manifest_file_list))
-
-    # END for
-
-    return game_data_list
-
-def get_matching_launcher_manifest(
-    game_manifest: FileDirectory,
-    launcher_manifest_file_list: list[FileDirectory]
-) -> FileDirectory:
-    
-    # Find launcher manifest that matches the name of the game manifest.
-    # Search result is None if match does not exist.
-    # Assumes there is only one or zero matching launcher manifests.
-    matching_launcher_manifest: FileDirectory = next(
-        (launcher_manifest for launcher_manifest in launcher_manifest_file_list if game_manifest.get_name_raw() == launcher_manifest.get_name_raw()),
-        None # Default
-    )
-
-    return matching_launcher_manifest
-
-def assert_manifest_is_supported(format_version: int):
-
-    if format_version not in SUPPORTED_LAUNCHER_MANIFEST_VERSIONS:
-
-        output: str = "ERROR!: Launcher manifest format version is incompatible.\n"
-        output += "Check if the new launcher \".item\" manifest format is compatible with this program.\n"
-        output += "Then add format version to SUPPORTED_LAUNCHER_MANIFEST_VERSIONS." 
-
-        print(output)
-        sys.exit(1)
 
 def update_manifest_location_references(launcher_manifest: FileDirectory, game_folder_path: str) -> None:
 
@@ -152,7 +34,7 @@ def update_manifest_location_references(launcher_manifest: FileDirectory, game_f
         data = json.load(file)
 
         # Check version
-        assert_manifest_is_supported(data["FormatVersion"])
+        GameDataManager.assert_manifest_is_supported(data["FormatVersion"])
 
         manifest_location = os.path.join(game_folder_path, GAME_MANIFEST_FOLDER_NAME)
         staging_location = os.path.join(manifest_location, STAGING_FOLDER_NAME)
@@ -175,7 +57,7 @@ def backup_game_manifest_files(
     for game_data in game_data_list:
         for game_manifest in game_data.manifest_file_list:
             
-            matching_launcher_manifest: FileDirectory = get_matching_launcher_manifest(game_manifest, launcher_manifest_file_list)
+            matching_launcher_manifest: FileDirectory = GameDataManager.get_matching_launcher_manifest(game_manifest, launcher_manifest_file_list)
 
             if matching_launcher_manifest is not None:
                 print(f"INFO: Backing up \"{matching_launcher_manifest.name}\".")
@@ -185,7 +67,7 @@ def backup_game_manifest_files(
 
 def restore_manifests(manifest_backup_folder: str, launcher_manifest_folder: str) -> None:
 
-    assert_path_exists(manifest_backup_folder, hint="You may need to backup manifests first.")
+    FileManagement.assert_path_exists(manifest_backup_folder, hint="You may need to backup manifests first.")
 
     if yes_no_prompt(f"Launcher manifests will restore to \"{launcher_manifest_folder}\". Continue?") == False:
         print("INFO: Aborting...")
@@ -194,7 +76,7 @@ def restore_manifests(manifest_backup_folder: str, launcher_manifest_folder: str
     print_line_separator()
 
     for manifest_entry in os.scandir(manifest_backup_folder):
-        if is_valid_launcher_manifest_file(manifest_entry):
+        if GameData.is_valid_launcher_manifest_file(manifest_entry):
             print(f"INFO: Restoring launcher manifest: {manifest_entry.name}")
             shutil.copy2(manifest_entry.path, launcher_manifest_folder)
     
@@ -208,9 +90,9 @@ def backup_manifests(manifest_backup_folder: str, launcher_manifest_folder: str,
 
     print_line_separator()
 
-    try_create_dir(manifest_backup_folder)
+    FileManagement.try_create_dir(manifest_backup_folder)
 
-    launcher_manifest_file_list: list[FileDirectory] = get_launcher_manifest_files(launcher_manifest_folder)
+    launcher_manifest_file_list: list[FileDirectory] = GameDataManager.get_launcher_manifest_files(launcher_manifest_folder)
 
     backup_game_manifest_files(manifest_backup_folder, launcher_manifest_file_list, game_data_list)
 
@@ -218,7 +100,7 @@ def backup_manifests(manifest_backup_folder: str, launcher_manifest_folder: str,
 
 def move_game_installation(manifest_backup_folder: str, game_data_list: list[GameData]) -> None:
     
-    assert_path_exists(manifest_backup_folder, hint="You may need to backup manifests first.")
+    FileManagement.assert_path_exists(manifest_backup_folder, hint="You may need to backup manifests first.")
 
     print("Movable Games Menu:")
 
@@ -259,7 +141,7 @@ def move_game_installation(manifest_backup_folder: str, game_data_list: list[Gam
     print_line_separator()    
 
     destination_path: str = input("Input a destination path: ")
-    assert_path_exists(destination_path)
+    FileManagement.assert_path_exists(destination_path)
     
     print_line_separator()
 
@@ -276,9 +158,9 @@ def move_game_installation(manifest_backup_folder: str, game_data_list: list[Gam
 
     # Create manifest backups folder in destination folder.
     destination_backup_folder = os.path.join(destination_path, MANIFEST_BACKUP_FOLDER_NAME)
-    try_create_dir(destination_backup_folder)
+    FileManagement.try_create_dir(destination_backup_folder)
 
-    launcher_manifest_file_list: list[FileDirectory] = get_launcher_manifest_files(manifest_backup_folder)
+    launcher_manifest_file_list: list[FileDirectory] = GameDataManager.get_launcher_manifest_files(manifest_backup_folder)
 
     for selected_game in selected_games_list:
 
@@ -286,7 +168,7 @@ def move_game_installation(manifest_backup_folder: str, game_data_list: list[Gam
 
         # Find matching launcher manifests within the backups folder.
         for game_manifest in selected_game.manifest_file_list:
-            matching_launcher_manifest = get_matching_launcher_manifest(game_manifest, launcher_manifest_file_list)
+            matching_launcher_manifest = GameDataManager.get_matching_launcher_manifest(game_manifest, launcher_manifest_file_list)
 
             if matching_launcher_manifest is None:
                 found_all_manifests = False
@@ -319,11 +201,11 @@ def relink_manifests(manifest_backup_folder: str, game_data_list: list[GameData]
         print("INFO: Aborting...")
         sys.exit(0)
 
-    launcher_manifest_file_list: list[FileDirectory] = get_launcher_manifest_files(manifest_backup_folder)
+    launcher_manifest_file_list: list[FileDirectory] = GameDataManager.get_launcher_manifest_files(manifest_backup_folder)
 
     for game_data in game_data_list:
         for game_manifest in game_data.manifest_file_list:
-            matching_launcher_manifest: FileDirectory = get_matching_launcher_manifest(game_manifest, launcher_manifest_file_list)
+            matching_launcher_manifest: FileDirectory = GameDataManager.get_matching_launcher_manifest(game_manifest, launcher_manifest_file_list)
 
             if matching_launcher_manifest is None:
                 print(f"WARNING!: Launcher manifest for \"{game_data.game_folder.name}\" matching {game_manifest.name} does not exist.")
@@ -349,17 +231,17 @@ def main():
     else:
         launcher_manifest_folder = input("Please input a valid path: ")
 
-    assert_path_exists(launcher_manifest_folder)
+    FileManagement.assert_path_exists(launcher_manifest_folder)
 
     # Get game data path ----------------------------------------------------
     
     game_data_path: str = input("Please enter your games folder path: ")
-    assert_path_exists(game_data_path)
+    FileManagement.assert_path_exists(game_data_path)
 
     # Get game folder list and game manifest list ----------------------------
 
     print_line_separator()
-    game_data_list: list[GameData] = get_game_data_list(game_data_path)
+    game_data_list: list[GameData] = GameDataManager.get_game_data_list(game_data_path)
     print_line_separator()
 
     if len(game_data_list) == 0:
